Moved more getValue and onTransition to the factory
