+++ src/microstates.ts
@@ -1,10 +1,9 @@
-import { __ } from 'ramda';
 import { IMicrostate, ISchema } from './Interfaces';
 import TypeTree from './utils/TypeTree';
 import mapState from './utils/mapState';
 import mapTransitions from './utils/mapTransitions';
-import getValue from './utils/getValue';
-import onTransitionFactory from './utils/onTransitionFactory';
+import getValue from './utils/getValueFactory';
+import onTransition from './utils/onTransitionFactory';
 
 export default function microstates(Class: ISchema, initial: any = undefined): IMicrostate {
   if (!(typeof Class === 'function' || Array.isArray(Class))) {
@@ -15,12 +14,15 @@ export default function microstates(Class: ISchema, initial: any = undefined): I
 
   let tree = new TypeTree(Class);
 
-  let onTransition = onTransitionFactory(function onTransitionHandler(newState: any) {
-    return mapState(tree, [], getValue(__, __, newState));
-  });
+  let state = mapState(tree, [], getValue(initial));
 
-  let state = mapState(tree, [], getValue(__, __, initial));
-  let transitions = mapTransitions(tree, [], onTransition(__, __, state));
+  let transitions = mapTransitions(
+    tree,
+    [],
+    onTransition(function onTransitionHandler(newState: any) {
+      return mapState(tree, [], getValue(newState));
+    }, state)
+  );
 
   return {
     state,
